updated tests
