+++ test/features/search.feature
@@ -3,8 +3,10 @@ Feature: Set of scenarios to test functionality of search filters on the search 
 
   Background:
     Given I have logged in
+    And a case type containing every field type exists
 
-    @broken
+
+  @broken
   Scenario: selecting the jurisdiction search drop down changes the main banner title to that jurisdiction name
     Given I am on the search page
     And the banner title matches that of the currently selected jurisdiction
@@ -12,8 +14,7 @@ Feature: Set of scenarios to test functionality of search filters on the search 
     Then the banner title matches that of the currently selected jurisdiction
 
   Scenario Outline: dynamic search filters are displayed for all data types
-    Given a case type containing every field type exists
-    And I am on the search page
+    Given I am on the search page
     When I select the 'Case type' drop down option for dynamic filters
     Then I should see a '<dataType>' dynamic filter with '<labels>' labels and '<values>' values
 
@@ -35,8 +36,7 @@ Feature: Set of scenarios to test functionality of search filters on the search 
 
     @searchtest
   Scenario Outline: reset button clears drop down options and removes all dynamic filters
-    Given a case type containing every field type exists
-    And I am on the search page
+    Given I am on the search page
     And I have filled out the search filters including dynamic filters
     When I click the 'Reset' button
     Then I should not see a '<dataType>' dynamic filter
@@ -64,12 +64,42 @@ Feature: Set of scenarios to test functionality of search filters on the search 
       | Text        |
 
   Scenario: apply button submits search options and returns results list
-    Given a case type containing every field type exists
-    And a case exists
+    Given a case exists
     And I am on the search page
     And I have filled out the search filters
     When I click the 'Apply' button
     Then the search result table will be displayed
+
+
+  @searchCaseList
+  Scenario: loading the results table will display correct result count
+    Given I am on the search page
+    When I have performed a search
+    Then I see cases listed in the results table
+    And I see table header with correct result count
+
+  @searchCaseList
+  Scenario: click on sorting icon on the first column will sort the results in descending order
+    Given I am on the search page
+    And I have some cases listed in the results table
+    When I click on sorting icon on the first column
+    Then the results of the page should be sorted in the descending order
+
+  @searchCaseList
+  Scenario: click on the case link will navigate to the details page
+    Given I am on the search page
+    And I have some cases listed in the results table
+    When I click on the case link
+    Then I see case details page
+
+  @searchCaseList
+  Scenario: loading results with more than 25 items will display at least 2 pages
+    Given I am on the search page
+    And I have more than 25 results
+    When I click on page link 2
+    Then I see results of the second page
+    And I see table header with page 2 result count
+    And The '2' is selected on the Pagination control
 
 #   @stickysearch
 #   Scenario: navigating away from search and back again via browser back saves search filters inputs
