+++ test/pageObjects/ccd-components/caseListComponent.js
@@ -11,6 +11,11 @@ class CaseListComponent extends BasePage {
     this._columnNamesList = '#search-result table thead table tr > div:nth-of-type(1)';
 
     this._columnDataCssTemplate = '#search-result table tbody tr td:nth-of-type(X)'
+    this._resultCount = by.css('.pagination-top');
+    this._sortIcon = by.css('.search-result-column-sort .sort-widget');
+    this._firstLink = by.css('#search-result table tbody tr:nth-of-type(1) td:nth-of-type(1) a');
+    this._pageTwoLink = by.css('.ngx-pagination li.current+li a');
+    this._selectedPaginationControlNumber = 'pagination-template li.current span:nth-of-type(2)';
   }
 
   /**
@@ -53,6 +58,56 @@ class CaseListComponent extends BasePage {
   async getFirstColumnResults (){
       let columnDataCss = this._columnDataCssTemplate.replace('X', 1);
       return await $$(columnDataCss)
+  }
+
+  /**
+   * Click Sort icon to sort results
+   * @returns {Promise<void|*>}
+   */
+  async clickSortIcon() {
+    await element(this._sortIcon).click();
+  }
+
+  /**
+   * Click Sort icon to sort results
+   * @returns {Promise<void|*>}
+   */
+  async clickFirstColumnResultLink() {
+    await element(this._firstLink).click();
+  }
+
+  /**
+   * Click link "2" on pagination area
+   * @returns {Promise<void|*>}
+   */
+  async clickPageTwo() {
+    await element(this._pageTwoLink).click();
+  }
+
+  /**
+   * get text in first line
+   * @returns {Promise<string|*>}
+   */
+  async clickFirstColumnResultText() {
+    return await element(this._firstLink).getText();
+  }
+
+  /**
+   * Return the result of total cases found above the case list table
+   * eg 'Displaying 1 - 25 out of 284 results'
+   * @returns {Promise<String|*>}
+   */
+  async getResultCountText(){
+    return await element(this._resultCount).getText();
+  }
+
+  /**
+   * Return the currently selected number selected on the Pagination control bar
+   * the bar is eg '<< Previous 1 2 3 ... n Next >>'
+   * @returns {Promise<void>}
+   */
+  async getSelectedPaginationControlNumber(){
+    return await $(this._selectedPaginationControlNumber).getText();
   }
 
 }
