+++ test/stepDefinitions/baseSteps.js
@@ -18,17 +18,6 @@ module.exports = {
     await wizardPage.clickSubmitCaseButton();
   },
 
-
-  fillOutAndSubmitFormWithValue: async function(value){
-    //todo eventually change to be dynamic and automatic
-    // This will create a case for 'All Field Data Types' caseType or any other case
-    //that has a optional text field and no other mandatory fields
-    let wizardPage = new CreateCaseWizardPage();
-    await wizardPage.interactWithField('text', value);
-    await wizardPage.clickContinueButton();
-    await wizardPage.clickSubmitCaseButton();
-
-  
   fillOutMandatoryFields: async function(){
     let wizardPage = new CreateCaseWizardPage();
     for (const elem of TestData.mandatoryFields) {
