+++ test/stepDefinitions/searchSteps.js
@@ -14,18 +14,15 @@ defineSupportCode(function ({ Given, When, Then, Before, After }) {
   Given(/^I am on the search page$/, async function () {
     await searchPage.getNavBarComponent().clickSearchLink();
     await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
-    // await searchPage.selectJurisdiction(TestData.jurisdiction);
   });
 
   Given(/^I have filled out the search filters$/, async function() {
     await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
-    // await searchPage.selectCaseType(TestData.caseType);
     await searchPage.interactWithField('text', 'text');
   });
 
   Given(/^I have filled out the search filters including dynamic filters$/, async function() {
     await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
-    // await searchPage.selectCaseType(TestData.caseType);
     await searchPage.interactWithField('text');
     await searchPage.interactWithField('textarea');
     await searchPage.interactWithField('number');
@@ -44,24 +41,19 @@ defineSupportCode(function ({ Given, When, Then, Before, After }) {
   When(/^I change the jurisdiction search drop down option$/, async function () {
     TestData.jurisdiction = 'Test';
     await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
-    // await searchPage.selectJurisdiction(TestData.jurisdiction);
   });
 
   When(/^I select the 'Case type' drop down option for dynamic filters$/, async function () {
     await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
-    // await searchPage.selectJurisdiction(TestData.jurisdiction);
     await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
-    // await searchPage.selectCaseType(TestData.caseType);
   });
 
   When(/^I click the 'Reset' button$/, async function () {
     await searchPage.getSearchFilters().clickResetButton();
-    // await searchPage.clickResetButton();
   });
 
   When(/^I click the 'Apply' button$/, async function () {
     await searchPage.getSearchFilters().clickApplyButton();
-    // await searchPage.clickApplyButton();
   });
 
   When(/^a case exists$/, async function () {
@@ -86,6 +78,114 @@ defineSupportCode(function ({ Given, When, Then, Before, After }) {
 
   Then(/^the search result table will be displayed$/, async function() {
     expect(await searchPage.isSearchResultsDisplayed()).to.be.true;
+  });
+
+  //Search Case List Steps
+  /**
+   * Checks search case list and continues to create a new case until there are 25 cases
+   * @returns {Promise<void>}
+   */
+  //todo broken?
+  async function bulkUpdateCases() {
+    let columnResults = await searchPage.getCaseListComponent().getFirstColumnResults();
+    if (columnResults.length < 25){
+        await baseSteps.navigateToCreateCasePage();
+        await baseSteps.fillOutAndSubmitForm();
+        await searchPage.getNavBarComponent().clickSearchLink();
+    }
+  }
+
+  Given(/^I have some cases listed in the results table$/, async function () {
+    await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
+    await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
+    await searchPage.getSearchFilters().clickApplyButton();
+
+    await bulkUpdateCases();
+  });
+
+  Given(/^I have more than 25 results$/, async function () {
+    await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
+    await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
+    await searchPage.getSearchFilters().clickApplyButton();
+    await bulkUpdateCases();
+  });
+
+  When(/^I have performed a search$/, async function () {
+    await searchPage.getSearchFilters().selectJurisdiction(TestData.jurisdiction);
+    await searchPage.getSearchFilters().selectCaseType(TestData.caseType);
+    await searchPage.getSearchFilters().clickApplyButton();
+    await bulkUpdateCases();
+  });
+
+  When(/^I click on sorting icon on the first column$/, async function () {
+    await searchPage.getCaseListComponent().clickSortIcon();
+  });
+
+  When(/^I click on the case link$/, async function () {
+    await searchPage.getCaseListComponent().clickFirstColumnResultLink();
+  });
+
+  When(/^I click on page link 2$/, async function () {
+    pageOneColumnOneFirstResult = await await searchPage.getCaseListComponent().clickFirstColumnResultText()
+    await searchPage.getCaseListComponent().clickPageTwo();
+  });
+
+  Then(/^I see cases listed in the results table$/, async function () {
+    let columnResults = await searchPage.getCaseListComponent().getFirstColumnResults();
+    expect (columnResults.length > 0).to.be.true;
+  });
+
+  Then(/^I see table header with correct result count$/, async function () {
+    const resultText = await searchPage.getCaseListComponent().getResultCountText();
+    const hasCorrectResult = resultText.includes('Displaying 1 - 25 out of ');
+
+    expect (hasCorrectResult).to.be.true;
+  });
+
+  Then(/^the results of the page should be sorted in the descending order$/, async function () {
+    const columnResults = await searchPage.getCaseListComponent().getFirstColumnResults();
+
+    let isSorted = true;
+    let tempArr = [];
+
+    for(let i = 0; i < columnResults.length - 1; i++) {
+      tempArr.push(await columnResults[i].getText());
+    }
+
+    //doesnt work well with text - 'aaa' 'AA' fails test - testing with case ref
+    for(let j = 0; j < tempArr.length - 1; j++) {
+      const firstElementText = tempArr[j];
+      const secondElementText = tempArr[j+1];
+      if(firstElementText < secondElementText) {
+        console.log(`first ${firstElementText}`)
+        console.log(`second ${secondElementText}`)
+        isSorted = false;
+        break;
+      }
+    }
+
+    expect (isSorted).to.be.true;
+  });
+
+  Then(/^I see case details page$/, async function () {
+    const url = await browser.getCurrentUrl();
+    expect(url).includes("/case/");
+  });
+
+  Then(/^I see results of the second page$/, async function () {
+    let pageTwoColumnOneFirstResult = await await searchPage.getCaseListComponent().clickFirstColumnResultText()
+    expect (pageTwoColumnOneFirstResult).to.not.equal(pageOneColumnOneFirstResult);
+  });
+
+  Then(/^I see table header with page 2 result count$/, async function () {
+    const resultText = await searchPage.getCaseListComponent().getResultCountText();
+    const hasCorrectResult = resultText.includes('Displaying 26');
+    expect (hasCorrectResult).to.be.true;
+  });
+
+  Then(/^The '2' is selected on the Pagination control/, async function () {
+    let selectedNumber = await searchPage.getCaseListComponent().getSelectedPaginationControlNumber();
+    expect(selectedNumber).to.eq('2')
   });
 
 });
